refactor: use semaphores for concurrency control
