+++ Cargo.toml
@@ -36,6 +36,7 @@ chrono = "0.4"
 thiserror = "2"
 serde = {version = "1", features = ["derive"]}
 csv = "1"
+dashmap = "6"
 
 # credentials
 keyring = "2"
