+++ src/cmd/config/mod.rs
@@ -4,7 +4,7 @@ use dialoguer::Confirm;
 use keyring::Entry;
 
 use self::{
-    credentials::{get_client_credentials, HandleCredentials},
+    credentials::HandleCredentials,
     models::{ConfigAuthCommand, ConfigCryptoCommand},
 };
 
@@ -19,6 +19,12 @@ pub mod logs;
 pub mod models;
 
 pub const MAX_CONCURRENT_REQUESTS: usize = 10;
+pub const DEFAULT_CHUNK_SIZE: usize = 1024 * 1024 * 32; // 32 MB (standard S3 chunk)
+pub const DEFAULT_CONCURRENT_MULTIPLIER: u8 = 10;
+pub const MAX_VELOCITY: u8 = 10;
+pub const MIN_VELOCITY: u8 = 1;
+pub const CLIENT_ID: &str = env!("DCCMD_CLIENT_ID");
+pub const CLIENT_SECRET: &str = env!("DCCMD_CLIENT_SECRET");
 
 pub struct ConfigCommandHandler {
     entry: Box<dyn HandleCredentials>,
@@ -34,7 +40,6 @@ impl ConfigCommandHandler {
     }
 
     async fn get_dracoon_client(&self, target: &str) -> Result<Dracoon<Connected>, DcCmdError> {
-        let (client_id, client_secret) = get_client_credentials();
         let Ok(refresh_token) = self.entry.get_dracoon_env() else {
             let msg = format_error_message(
                 format!("No token found for this DRACOON url: {target}.").as_str(),
@@ -47,8 +52,8 @@ impl ConfigCommandHandler {
 
         let dracoon = Dracoon::builder()
             .with_base_url(target)
-            .with_client_id(client_id)
-            .with_client_secret(client_secret)
+            .with_client_id(CLIENT_ID)
+            .with_client_secret(CLIENT_SECRET)
             .build()?
             .connect(OAuth2Flow::refresh_token(refresh_token))
             .await?;
