+++ src/cmd/nodes/mod.rs
@@ -2,10 +2,7 @@ use std::sync::Arc;
 
 use console::Term;
 use dialoguer::Confirm;
-use futures_util::{
-    future::{join_all, try_join_all},
-    stream, StreamExt,
-};
+use futures_util::{future::try_join_all, stream, StreamExt};
 use models::{CmdCopyOptions, CmdListNodesOptions};
 use tokio::sync::Mutex;
 use tracing::{debug, error, info};
@@ -29,6 +26,7 @@ use dco3::{
 use self::models::CmdMkRoomOptions;
 
 use super::{
+    config::MAX_CONCURRENT_REQUESTS,
     models::{build_params, DcCmdError, ListOptions, PasswordAuth},
     utils::strings::{format_error_message, format_success_message},
 };
@@ -115,29 +113,58 @@ async fn get_nodes(
         .await?;
 
     if opts.all() && node_list.range.total > 500 {
-        let mut offset = 500;
-        let limit = 500;
+        let (tx, mut rx) = tokio::sync::mpsc::channel(MAX_CONCURRENT_REQUESTS);
+        let semaphore = Arc::new(tokio::sync::Semaphore::new(MAX_CONCURRENT_REQUESTS));
+        let mut handles = Vec::new();
 
-        while offset < node_list.range.total {
-            let mut futures = vec![];
+        (500..=node_list.range.total)
+            .step_by(500)
+            .for_each(|offset| {
+                let tx = tx.clone();
+                let semaphore = semaphore.clone();
+                let dracoon = dracoon.clone();
+                let opts = opts.clone();
 
-            while offset < node_list.range.total {
-                let params = build_params(opts.filter(), offset, None)?;
+                let handle = tokio::spawn(async move {
+                    let _permit = semaphore.acquire().await.map_err({
+                        error!("Failed to acquire semaphore permit.");
+                        |_| DcCmdError::IoError
+                    })?;
+                    let params = build_params(opts.filter(), offset, None)?;
 
-                let next_node_list_req =
-                    dracoon.nodes().get_nodes(parent_id, managed, Some(params));
-                futures.push(next_node_list_req);
-                offset += limit;
+                    match dracoon
+                        .nodes()
+                        .get_nodes(parent_id, managed, Some(params))
+                        .await
+                    {
+                        Ok(node_list) => {
+                            if let Err(e) = tx.send(node_list.items).await {
+                                error!("Failed to send node list: {}", e);
+                                return Err(DcCmdError::IoError);
+                            }
+
+                            Ok::<(), DcCmdError>(())
+                        }
+                        Err(e) => {
+                            error!("Error getting folders: {}", e);
+                            Err(e.into())
+                        }
+                    }
+                });
+                handles.push(handle);
+            });
+
+        drop(tx);
+
+        while let Some(result) = rx.recv().await {
+            node_list.items.extend(result);
+        }
+
+        for handle in handles {
+            if let Err(e) = handle.await {
+                error!("Failed to join task: {}", e);
+                return Err(DcCmdError::IoError);
             }
-
-            let mut next_node_list_items = vec![];
-
-            let results = join_all(futures).await;
-            for result in results {
-                let next_node_list = result?.items;
-                next_node_list_items.extend(next_node_list);
-            }
-            node_list.items.append(&mut next_node_list_items);
         }
     }
     Ok(node_list)
@@ -179,31 +206,59 @@ async fn search_nodes(
         .await?;
 
     if opts.all() && node_list.range.total > 500 {
-        let mut offset = 500;
-        let limit = 500;
+        let (tx, mut rx) = tokio::sync::mpsc::channel(MAX_CONCURRENT_REQUESTS);
+        let semaphore = Arc::new(tokio::sync::Semaphore::new(MAX_CONCURRENT_REQUESTS));
+        let mut handles = Vec::new();
 
-        while offset < node_list.range.total {
-            let mut futures = vec![];
+        (500..=node_list.range.total)
+            .step_by(500)
+            .for_each(|offset| {
+                let tx = tx.clone();
+                let semaphore = semaphore.clone();
+                let dracoon = dracoon.clone();
+                let opts = opts.clone();
+                let search_string = search_string.to_string();
 
-            while offset < node_list.range.total {
-                let params = build_params(opts.filter(), offset, None)?;
+                let handle = tokio::spawn(async move {
+                    let _permit = semaphore.acquire().await.map_err({
+                        error!("Failed to acquire semaphore permit.");
+                        |_| DcCmdError::IoError
+                    })?;
+                    let params = build_params(opts.filter(), offset, None)?;
 
-                let next_node_list_req =
-                    dracoon
+                    match dracoon
                         .nodes()
-                        .search_nodes(search_string, parent_id, Some(0), Some(params));
-                futures.push(next_node_list_req);
-                offset += limit;
-            }
+                        .search_nodes(&search_string, parent_id, Some(0), Some(params))
+                        .await
+                    {
+                        Ok(node_list) => {
+                            if let Err(e) = tx.send(node_list.items).await {
+                                error!("Failed to send node list: {}", e);
+                                return Err(DcCmdError::IoError);
+                            }
 
-            let mut next_node_list_items = vec![];
+                            Ok::<(), DcCmdError>(())
+                        }
+                        Err(e) => {
+                            error!("Error getting folders: {}", e);
+                            Err(e.into())
+                        }
+                    }
+                });
+                handles.push(handle);
+            });
 
-            let results = join_all(futures).await;
-            for result in results {
-                let next_node_list = result?.items;
-                next_node_list_items.extend(next_node_list);
+        drop(tx);
+
+        while let Some(result) = rx.recv().await {
+            node_list.items.extend(result);
+        }
+
+        for handle in handles {
+            if let Err(e) = handle.await {
+                error!("Failed to join task: {}", e);
+                return Err(DcCmdError::IoError);
             }
-            node_list.items.append(&mut next_node_list_items);
         }
     }
     Ok(node_list)
